Function ic_crypto_utils_threshold_sig::verify_combined

pub fn verify_combined<T: Signable>(
    msg: &T, 
    sig: &CombinedThresholdSigOf<T>, 
    pk: &ThresholdSigPublicKey
) -> CryptoResult<()>

Verify a combined threshold signature.

Arguments

msg is the Signable object associated with the signature

sig is the combined threshold signature to be verified

pk is the public key

Returns

Ok(()) if the signature is accepted, or an Err otherwise

Error

Returns an error if the signature could not be verified